[#1326] Operator: extend Agreements service with context
