+++ bundles/org.eclipse.passage.loc.licenses.emfforms/viewmodels/LicensePlan.view
@@ -1,22 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <org.eclipse.emf.ecp.view.model:View xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore" xmlns:org.eclipse.emf.ecp.view.model="http://org/eclipse/emf/ecp/view/model/1180" xmi:id="_PPv1MIdmEemHgffLhDrdUQ" name="License Plan">
   <rootEClass href="http://www.eclipse.org/passage/lic/licenses/2.0.0#//LicensePlan"/>
-  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_PP2i4IdmEemHgffLhDrdUQ" name="Control identifier">
+  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_PP2i4IdmEemHgffLhDrdUQ" name="Identifier">
     <domainModelReference xsi:type="org.eclipse.emf.ecp.view.model:FeaturePathDomainModelReference" xmi:id="_PP2i4YdmEemHgffLhDrdUQ">
       <domainModelEFeature xsi:type="ecore:EAttribute" href="http://www.eclipse.org/passage/lic/licenses/2.0.0#//LicensePlan/identifier"/>
     </domainModelReference>
   </children>
-  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_PP2i4odmEemHgffLhDrdUQ" name="Control name">
+  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_PP2i4odmEemHgffLhDrdUQ" name="Name">
     <domainModelReference xsi:type="org.eclipse.emf.ecp.view.model:FeaturePathDomainModelReference" xmi:id="_PP2i44dmEemHgffLhDrdUQ">
       <domainModelEFeature xsi:type="ecore:EAttribute" href="http://www.eclipse.org/passage/lic/licenses/2.0.0#//LicensePlan/name"/>
     </domainModelReference>
   </children>
-  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_PP2i5IdmEemHgffLhDrdUQ" name="Control description">
+  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_PP2i5IdmEemHgffLhDrdUQ" name="Description">
     <domainModelReference xsi:type="org.eclipse.emf.ecp.view.model:FeaturePathDomainModelReference" xmi:id="_PP2i5YdmEemHgffLhDrdUQ">
       <domainModelEFeature xsi:type="ecore:EAttribute" href="http://www.eclipse.org/passage/lic/licenses/2.0.0#//LicensePlan/description"/>
     </domainModelReference>
   </children>
-  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_9cX1gP6MEeuGytDyYvaNeQ" name="agreements">
+  <children xsi:type="org.eclipse.emf.ecp.view.model:Control" xmi:id="_9cX1gP6MEeuGytDyYvaNeQ" name="Agreements">
     <domainModelReference xsi:type="org.eclipse.emf.ecp.view.model:FeaturePathDomainModelReference" xmi:id="_DSQq0P6NEeuGytDyYvaNeQ">
       <domainModelEFeature xsi:type="ecore:EAttribute" href="http://www.eclipse.org/passage/lic/licenses/2.0.0#//LicensePlan/agreements"/>
     </domainModelReference>
