+++ bundles/org.eclipse.passage.loc.workspace/src/org/eclipse/passage/loc/workspace/LocalFileHandle.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 ArSysOp
+ * Copyright (c) 2021, 2024 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -9,6 +9,7 @@
  *
  * Contributors:
  *     ArSysOp - initial API and implementation
+ *     ArSysOp - further support
  *******************************************************************************/
 package org.eclipse.passage.loc.workspace;
 
@@ -23,10 +24,6 @@ import org.eclipse.emf.common.util.URI;
 import org.eclipse.passage.loc.internal.api.workspace.ResourceHandle;
 import org.eclipse.passage.loc.internal.api.workspace.ResourceType;
 
-/**
- * temporary
- * 
- */
 class LocalFileHandle implements ResourceHandle {
 
 	private final ResourceType type;
